# docs/swagger.yaml
basePath: /api/v1
definitions:
  categorya_dto.CategoryResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        additionalProperties:
          type: string
        type: object
      updated_at:
        type: string
    type: object
  categorya_dto.CreateCategoryRequest:
    properties:
      is_active:
        type: boolean
      name:
        additionalProperties:
          type: string
        type: object
    required:
    - name
    type: object
  categorya_dto.UpdateCategoryRequest:
    properties:
      is_active:
        type: boolean
      name:
        additionalProperties:
          type: string
        type: object
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Admin paneli uchun CRUD endpointlari (cursor pagination, filter, sort)
  title: Usta Backend Admin API
  version: "1.0"
paths:
...
  /categories:
    get:
      parameters:
      - description: Keyset cursor
        in: query
        name: cursor
        type: string
      - description: Default 10, max 100
        in: query
        name: limit
        type: integer
      - description: id|name|is_active|created_at|updated_at
        in: query
        name: sort_by
        type: string
      - description: asc|desc
        in: query
        name: sort_order
        type: string
      - description: Name JSONB ichida ILIKE
        in: query
        name: name
        type: string
      - description: Holat
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Kategoriyalar ro'yxati (admin, cursor pagination)
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: 'name JSONB: har bir faol til kalitlari kiritilishi kerak'
      parameters:
      - description: Kategoriya ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/categorya_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/categorya_dto.CategoryResponse'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Kategoriya yaratish (admin, ko'p tilli)
      tags:
      - Categories
  /categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Kategoriyani o'chirish (soft delete, admin)
      tags:
      - Categories
    get:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categorya_dto.CategoryResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Kategoriyani ID bo'yicha olish (admin)
      tags:
      - Categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Yangilanadigan maydonlar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/categorya_dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categorya_dto.CategoryResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Kategoriyani yangilash (admin)
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"